reference-app/backend/requirements.txt
Flask==2.1.0
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
prometheus-client==0.12.0
prometheus-flask-exporter
jaeger-client
opentelemetry-instrumentation
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-flask
opentelemetry-api